==> fisheries/marine/disturbance/STEP2b_classify_gear_type.Rmd <==
---
title: "Classifying Gear Types"
author: "Gage Clawson (UCSB, NCEAS, OHI)"
date: "October 2, 2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

This script assigns gear and species specific data to the raw data from [Watson v5.0 (2018)](https://www.nature.com/articles/sdata201739) in order to classify catch as either **demersal** or **pelagic**, **destructive** or **nondestructive** and **high** or **low** bycatch. Using the gear classifications laid out in Halpern et al. (2008) Table S4, as well as species information from [FishBase](http://www.fishbase.org/search.php) shared by the [Sea Around Us Project](http://www.seaaroundus.org/) data we were able to assign gear classifications to each record.

Overview:

- Each data file provided by Watson was provided by Watson via email, and is held on a server held at NCEAS.
- After extracting all unique gear types found in the data, a *gear to category* dataset was created (`gear_to_cat.csv`) manually using information from Table S4 in Halpern et al. (2008), and joined to the raw data. This dataset added three new columns to the data; **bycatch** (high or low), **type** (pelagic or demersal), and **destruction** (destructive or nondestructive). 
- The Watson v5.0 data has information on **Functional Group** for each species, which was used to assign either *pelagic* or *demersal* to species for the *gear_to_cat.csv* data.
- Remaining catch records with missing or mismatched information were cleaned by hand which is detailed in this script (there were none).
- Final output is a dataset, `taxa_gear_types.csv` that assigns each taxa+gear type to the categories used to create the five fishing stressors.


# Setup
```{r setup,message=F,warning=F}
knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',message = FALSE, warning = FALSE)
source('https://raw.githubusercontent.com/OHI-Science/ohiprep_v2020/f2f4f0592d29e465dba2b3889d110cbda56eeff9/workflow/R/common.R')
library(raster)
library(readr)
library(dplyr)
library(data.table)
library(stringr)
```

# Get gear types from raw data

Read in one of the raw data files from Watson v5.0 (2018) and look at the unique gear types.

```{r raw_data_files, eval = FALSE}

## read in gear index 
raw <- read_csv(file.path("/home/shares/food-systems/Food_footprint/_raw_data/watson_data/v5.0/Codes_gear.csv"))
#look at all the unique gear types
gear <- raw %>%
          dplyr::select(Gear,VBDesc)%>%
          unique()
DT::datatable(gear,rownames = F)

write.csv(gear, "fisheries/marine/disturbance/int/gears.csv", row.names = FALSE)
```


The bycatch and destruction information was manually added for each gear type using figure S4 from Halpern (2008). I have not filled out the type of fish yet, demersal or pelagic. I will use the Descript column from the catch dataset to determine the type. This information was saved in `gear_to_cat.csv`.


```{r, eval = FALSE}
gear_cat <- read_csv('fisheries/marine/disturbance/int/gears_to_cat.csv') %>%
  dplyr::select(Gear, GearName = VBDesc, bycatch, destruction)
DT::datatable(gear_cat,rownames = F)
```

# Assign gear categories to data

Since we are just focusing on assigning gear types to the catch records we can eliminate some of the columns from the raw dataset and select unique records for Taxonkey, TaxonName, CommonName, Gear, GearName.

```{r, eval = FALSE}

## read in raw watson catch data 
raw_catch <- read_csv(file.path("/home/shares/food-systems/Food_footprint/_raw_data/watson_data/v5.0/Catch2015_2019.csv"))

## read in taxa index
raw_taxa <- read_csv(file.path("/home/shares/food-systems/Food_footprint/_raw_data/watson_data/v5.0/Codes_taxa.csv"))

## join the data, checking the rows to make sure there are no row duplications
data <- raw_catch %>%
        left_join(raw_taxa, by = "Taxonkey") %>%
        left_join(gear, by = "Gear") %>% 
        dplyr::select(Taxonkey,TaxonName,CommonName,Gear,GearName = VBDesc, Descript)%>%
        unique()%>%
        left_join(gear_cat,by = c("Gear","GearName"))
DT::datatable(head(data,n=50),rownames = F)
```

# Assign species type categories

We still have a lot of records that need further clarification with regards to type of species (demersal or pelagic) since the GearName itself does not distinguish between these two. We can use the "Descript" column from the taxa dataset provided by Watson. 


```{r, eval = FALSE}
catch_df_taxa <-data%>%
                dplyr::select(Taxonkey, TaxonName,CommonName,Descript)%>%
                unique()

nas <- filter(catch_df_taxa,is.na(Descript))
```

Sticking with the `catch_df_taxa dataset`, lets create new columns that tease out the type of species either demersal or pelagic using the `stringr` package function, `str_detect`. Here we assign the *pelagic* category to all records which have a Group Description containing: "reef", "pelagic", "Krill", and "Cephalopods". We assign the *demersal* category to all records that contain: "Shrimp", "crab","lobster","demersal", "shark", "ray",or "flatfish" in the description.
.

```{r, eval = FALSE}
## classify species types into demersal or pelagic
catch_df_taxa <- catch_df_taxa %>%
                  mutate(type = ifelse(str_detect(Descript, "benthopelagic|shrimp|crab|lobster|demersal|flatfish|shark|ray"),"demersal", ifelse(str_detect(Descript, "reef|pelagic >|pelagic <|pelagic 30|krill|cephalopods"), "pelagic", NA)))
DT::datatable(catch_df_taxa,rownames = F)

test <- catch_df_taxa %>%
  dplyr::filter(str_detect(Descript, "benthopelagic"))

test <- catch_df_taxa %>%
  dplyr::filter(is.na(type))

unique(catch_df_taxa$Descript)
```

Add taxonomic info back to the dataset. Now we have two different ways of categorizing each species, the first was by gear type and the second by functional group. We need to check and make sure these line up.

```{r, eval = FALSE}
data_plus_gear_info <- data %>%
                        left_join(catch_df_taxa,by = c('Taxonkey','CommonName','TaxonName', 'Descript')) 
```


We have assigned each gear & species combination to either the *pelagic* or *demersal* category as well as destruction (non-destructive or destructive) but there are still gear types that do not have a bycatch assigned, specifically two types of gear, **without gear** and **Lines**. Again we can rely on table S4 from Halpern et al. (2008) and assign **Lines**, which are all for pelagic species, to "low bycatch". We are also assigning **without gear** to be low bycatch since many of these are hand collection.

```{r, eval = FALSE}
DT::datatable(data_plus_gear_info %>%
                filter(is.na(bycatch)),rownames = F)


test <- data_plus_gear_info %>%
                filter(is.na(bycatch))
unique(test$GearName) # "without gear" "Lines"
```

The final dataset that matches each unique gear & taxon from the raw Watson data to the information we will use to map each of the 5 catch layers is saved as `taxa_gear_types.csv`

```{r assign_bycatch, eval=FALSE}
final_df <- data_plus_gear_info %>%
            mutate(bycatch = ifelse(is.na(bycatch),'low',bycatch))

summary(final_df)

write.csv(final_df,'fisheries/marine/disturbance/int/taxa_gear_types.csv', row.names = FALSE)
```
